Change indent, Add --force option.

+++ checkout_externals.py
@@ -1,26 +1,41 @@
+#!/usr/bin/python
+
 import sys
 import os
 import shutil
+import optparse
 
 vars = {}
 deps = {}
 deps_os = {}
 
 def Var(name):
-    return vars[name]
+  return vars[name]
 
-def checkout(dir, url):
+def checkout(dir, url, force=False):
+  if force:
     shutil.rmtree(dir, True)
-    os.system("svn co %s %s" % (url, dir))
+  os.system("svn checkout --trust-server-cert %s %s" % (url, dir))
+
+def parse_args():
+  parser = optparse.OptionParser(usage="%prog [options] {" + ", ".join(deps_os.keys()) + "}")
+  parser.add_option("-f", "--force", help="force update externals",
+                    action="store_true", dest="force", default=False)
+  return parser.parse_args()
 
 def main():
-    execfile('src/DEPS', globals(), globals())
+  execfile(os.path.join(os.path.dirname(__file__), "src", "DEPS"), globals(), globals())
 
-    plat = sys.argv[1]
+  (options, args) = parse_args()
+
+  if len(args) > 0 and deps_os.has_key(args[0]):
     for dir, url in deps.iteritems():
-        checkout(dir, url)
-    for dir, url in deps_os[plat].iteritems():
-        checkout(dir, url)
+      checkout(dir, url, options.force)
+    for dir, url in deps_os[args[0]].iteritems():
+      checkout(dir, url, options.force)
+  else:
+    print >> sys.stderr, "unknown or missing platform."
+    exit(1)
 
 if __name__ == '__main__':
-    main()
+  main()
